backend/auth_service/requirements/base.txt
# Core Django
Django==5.2.1
django-environ==0.12.0

# Database
psycopg2-binary==2.9.10

# Django REST Framework
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0

# JWT Authentication
PyJWT==2.8.0

# OAuth
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
requests-oauthlib==2.0.0

# Password hashing
argon2-cffi==23.1.0

# Utilities
python-decouple==3.8
whitenoise==6.9.0

# Email
django-anymail==12.0

# Caching
redis==5.0.8
django-redis==5.4.0

# Validation
cerberus==1.3.5

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Production server
gunicorn==23.0.0

# Monitoring
sentry-sdk==2.28.0
